+++ src/index.ts
@@ -1,5 +1,7 @@
 export { invariant } from './invariant';
 export { invariantAs } from './invariant-as';
 export { assertNever } from './assert-never';
+export { assertDefined } from './assert-defined';
+export { assertType } from './assert-type';
 export { unreachable } from './unreachable';
 export { InvariantError } from './types';